Add support for upstart for ubuntu customers.
Older versions of 14.04 still use upstart. This adds support for
upstart. On ubuntu with systemd.

service docker-vmdk-plugin <command> still uses systemd on newer
versions of ubuntu.

Photon (rpm) only uses systemd.

+++ scripts/install/systemd-before-remove.sh
@@ -14,5 +14,11 @@
 # limitations under the License.
 
 
-systemctl stop docker-vmdk-plugin.service
-systemctl disable docker-vmdk-plugin.service
+stat /proc/1/exe | grep File | grep systemd
+if [ $? -eq 0 ]
+then
+  systemctl stop docker-vmdk-plugin.service
+  systemctl disable docker-vmdk-plugin.service
+else
+  service docker-vmdk-plugin stop
+fi
